is applied

+++ example1.py
@@ -9,7 +9,7 @@ class Test(TorchPlus):
     def process(self):
         proc = self.input('input',torch.FloatTensor([2.,4.])) * self.parameter([1],'param')
 
-        self.label(torch.FloatTensor([18.,36.]))
+        self.label([18.,36.])
         return proc
 
 tp = Test()
